requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
chardet==3.0.4
idna==2.8
multidict==4.6.1
pyserial==3.4
pyserial-asyncio==0.4
python-engineio==3.11.0
python-socketio==4.4.0
six==1.13.0
websockets==8.1
yarl==1.4.2
